finish the importing of metric framework.
