--- metrics/graphite_test.go
@@ -1,22 +1,0 @@
-package metrics
-
-import (
-	"net"
-	"time"
-)
-
-func ExampleGraphite() {
-	addr, _ := net.ResolveTCPAddr("net", ":2003")
-	go Graphite(DefaultRegistry, 1*time.Second, "some.prefix", addr)
-}
-
-func ExampleGraphiteWithConfig() {
-	addr, _ := net.ResolveTCPAddr("net", ":2003")
-	go GraphiteWithConfig(GraphiteConfig{
-		Addr:          addr,
-		Registry:      DefaultRegistry,
-		FlushInterval: 1 * time.Second,
-		DurationUnit:  time.Millisecond,
-		Percentiles:   []float64{0.5, 0.75, 0.99, 0.999},
-	})
-}
