+++ metrics/metrics_test.go
@@ -25,7 +25,7 @@ func BenchmarkMetrics(b *testing.B) {
 	h := NewRegisteredHistogram("histogram", r, NewUniformSample(100))
 	m := NewRegisteredMeter("meter", r)
 	t := NewRegisteredTimer("timer", r)
-	RegisterDebugGCStats(r)
+	//RegisterDebugGCStats(r)
 	RegisterRuntimeMemStats(r)
 	b.ResetTimer()
 	ch := make(chan bool)
